network_socket_program/requirements.txt
pycryptodome==3.12.0
PyQt5==5.15.6
